requirements.txt

Django==3.1.7
cryptography==3.4.7
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
django-filter==2.4.0
Pillow==8.1.2
djoser==2.1.0
psycopg2-binary==2.8.6
django-filter==2.4.0
celery==5.0.5